# architecture/releases/v0.4/phase2-volume1/contracts/phase2_api_contract.yaml
openapi: 3.1.0
info:
  title: Sapientia Enterprise Intelligence Experience API
  version: 0.4.0
  description: Additive Phase 2 API contract. Existing APIs remain unchanged.
paths:
  /intelligence-experience/domains/{domain_code}/narrative:
    get:
      summary: Get the current structured domain narrative
      parameters:
        - $ref: '#/components/parameters/DomainCode'
        - $ref: '#/components/parameters/ProjectId'
        - name: assessment_id
          in: query
          required: false
          schema: { type: integer, minimum: 1 }
        - name: ai_mode
          in: query
          required: false
          schema:
            type: string
            enum: [deterministic, enriched]
            default: deterministic
      responses:
        '200':
          description: Domain narrative
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DomainNarrative' }
        '404': { description: Domain or assessment not found }
  /intelligence-experience/domains/{domain_code}/story:
    post:
      summary: Generate or refresh the consultant-style story
      parameters:
        - $ref: '#/components/parameters/DomainCode'
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                assessment_id: { type: integer, minimum: 1 }
                force_refresh: { type: boolean, default: false }
                tone: { type: string, enum: [executive, detailed], default: executive }
      responses:
        '200':
          description: Generated story
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DomainNarrative' }
        '503': { description: AI generation required but unavailable }
  /intelligence-experience/domains/{domain_code}/health:
    get:
      summary: Get explained business health
      parameters:
        - $ref: '#/components/parameters/DomainCode'
        - $ref: '#/components/parameters/ProjectId'
        - name: assessment_id
          in: query
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Business health
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BusinessHealth' }
  /intelligence-experience/domains/{domain_code}/timeline:
    get:
      summary: Get Phase 2 assessment evolution timeline
      parameters:
        - $ref: '#/components/parameters/DomainCode'
        - $ref: '#/components/parameters/ProjectId'
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 30 }
      responses:
        '200':
          description: Timeline
          content:
            application/json:
              schema:
                type: object
                required: [project_id, business_domain, timeline]
                properties:
                  project_id: { type: integer }
                  business_domain: { type: string }
                  timeline:
                    type: array
                    items: { $ref: '#/components/schemas/TimelineEvent' }
  /intelligence-experience/business-objects/{enterprise_object_id}/profile:
    get:
      summary: Get a business-first profile for an enterprise object
      parameters:
        - name: enterprise_object_id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
        - $ref: '#/components/parameters/ProjectId'
        - name: domain_code
          in: query
          required: true
          schema: { type: string }
        - name: assessment_id
          in: query
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Business object profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BusinessObjectProfile' }
        '404': { description: Business object not found }
  /intelligence-experience/statements/{statement_id}/explain:
    post:
      summary: Explain one narrative statement using its bounded evidence
      parameters:
        - name: statement_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Evidence-backed explanation
          content:
            application/json:
              schema:
                type: object
                required: [statement_id, explanation, support_status, evidence]
                properties:
                  statement_id: { type: string }
                  explanation: { type: string }
                  support_status: { $ref: '#/components/schemas/SupportStatus' }
                  confidence: { type: number, minimum: 0, maximum: 1 }
                  evidence:
                    type: array
                    items: { $ref: '#/components/schemas/EvidenceReference' }
components:
  parameters:
    DomainCode:
      name: domain_code
      in: path
      required: true
      schema: { type: string, minLength: 1 }
    ProjectId:
      name: project_id
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
  schemas:
    SupportStatus:
      type: string
      enum: [SUPPORTED, PARTIALLY_SUPPORTED, INSUFFICIENT_EVIDENCE]
    EvidenceReference:
      type: object
      required: [evidence_id, evidence_type, support_status]
      properties:
        evidence_id: { type: integer }
        evidence_type: { type: string }
        title: { type: string }
        excerpt: { type: [string, 'null'] }
        source: { type: [string, 'null'] }
        confidence: { type: [number, 'null'], minimum: 0, maximum: 1 }
        support_status: { $ref: '#/components/schemas/SupportStatus' }
        enterprise_object_id: { type: [integer, 'null'] }
        dataset_id: { type: [integer, 'null'] }
        column_id: { type: [integer, 'null'] }
        knowledge_item_id: { type: [integer, 'null'] }
    NarrativeStatement:
      type: object
      required: [statement_id, section, headline, text, support_status, generated_by, evidence]
      properties:
        statement_id: { type: string }
        section: { type: string }
        headline: { type: string }
        text: { type: string }
        support_status: { $ref: '#/components/schemas/SupportStatus' }
        confidence: { type: [number, 'null'], minimum: 0, maximum: 1 }
        generated_by: { type: string, enum: [DETERMINISTIC, AI_ENRICHED, AI_GENERATED] }
        intelligence_object_ids:
          type: array
          items: { type: integer }
        business_object_ids:
          type: array
          items: { type: integer }
        evidence:
          type: array
          items: { $ref: '#/components/schemas/EvidenceReference' }
    BusinessHealth:
      type: object
      required: [label, explanation, confidence, evidence_coverage, positive_drivers, negative_drivers]
      properties:
        label:
          type: string
          enum: [HEALTHY, STABLE, NEEDS_ATTENTION, CRITICAL, INSUFFICIENT_EVIDENCE]
        score: { type: [number, 'null'], minimum: 0, maximum: 100 }
        explanation: { type: string }
        confidence: { type: number, minimum: 0, maximum: 1 }
        evidence_coverage: { type: number, minimum: 0, maximum: 1 }
        positive_drivers:
          type: array
          items: { $ref: '#/components/schemas/NarrativeStatement' }
        negative_drivers:
          type: array
          items: { $ref: '#/components/schemas/NarrativeStatement' }
    EnterpriseIntelligenceObject:
      type: object
      required: [intelligence_object_id, assessment_id, object_type, object_key, title, status]
      properties:
        intelligence_object_id: { type: integer }
        assessment_id: { type: integer }
        object_type:
          type: string
          enum: [OBSERVATION, FINDING, RISK, OPPORTUNITY, KPI, RECOMMENDATION, ROOT_CAUSE, BUSINESS_IMPACT, TREND, ANOMALY, INSIGHT]
        object_key: { type: string }
        title: { type: string }
        description: { type: [string, 'null'] }
        interpretation: { type: [string, 'null'] }
        status: { type: string }
        category: { type: [string, 'null'] }
        severity: { type: [string, 'null'] }
        priority: { type: [string, 'null'] }
        confidence_score: { type: [number, 'null'], minimum: 0, maximum: 1 }
        probability_score: { type: [number, 'null'], minimum: 0, maximum: 1 }
        impact_score: { type: [number, 'null'], minimum: 0, maximum: 1 }
        enterprise_object_id: { type: [integer, 'null'] }
        evidence_count: { type: integer, minimum: 0 }
        attributes: { type: object, additionalProperties: true }
    DomainNarrative:
      type: object
      required: [experience_version, narrative_schema_version, project_id, business_domain, assessment, executive_summary, business_health, sections, provenance]
      properties:
        experience_version: { type: string, const: '0.4' }
        narrative_schema_version: { type: string, const: '1.0' }
        project_id: { type: integer }
        business_domain: { type: string }
        assessment:
          type: object
          required: [assessment_id, assessment_version, assessment_status, assessment_title, generated_at]
          properties:
            assessment_id: { type: integer }
            assessment_version: { type: integer }
            assessment_status: { type: string }
            assessment_title: { type: string }
            generated_at: { type: string, format: date-time }
            overall_confidence: { type: [number, 'null'], minimum: 0, maximum: 1 }
        executive_summary: { $ref: '#/components/schemas/NarrativeStatement' }
        business_health: { $ref: '#/components/schemas/BusinessHealth' }
        sections:
          type: object
          required: [current_state, what_changed, why_it_changed, risks, opportunities, recommendations]
          properties:
            current_state: { type: array, items: { $ref: '#/components/schemas/NarrativeStatement' } }
            what_changed: { type: array, items: { $ref: '#/components/schemas/NarrativeStatement' } }
            why_it_changed: { type: array, items: { $ref: '#/components/schemas/NarrativeStatement' } }
            risks: { type: array, items: { $ref: '#/components/schemas/EnterpriseIntelligenceObject' } }
            opportunities: { type: array, items: { $ref: '#/components/schemas/EnterpriseIntelligenceObject' } }
            recommendations: { type: array, items: { $ref: '#/components/schemas/EnterpriseIntelligenceObject' } }
        comparison_available: { type: boolean }
        suggested_questions: { type: array, items: { type: string } }
        provenance:
          type: object
          required: [generated_at, generation_mode, deterministic_fallback, source_object_count, evidence_count]
          properties:
            generated_at: { type: string, format: date-time }
            generation_mode: { type: string, enum: [DETERMINISTIC, AI_ENRICHED, AI_GENERATED] }
            deterministic_fallback: { type: boolean }
            prompt_version: { type: [string, 'null'] }
            model_provider: { type: [string, 'null'] }
            model_name: { type: [string, 'null'] }
            source_object_count: { type: integer }
            evidence_count: { type: integer }
    TimelineEvent:
      type: object
      required: [event_id, event_type, occurred_at, title]
      properties:
        event_id: { type: string }
        event_type: { type: string, enum: [ASSESSMENT_PUBLISHED, OBJECT_NEW, OBJECT_CHANGED, OBJECT_RESOLVED, RECOMMENDATION_CREATED] }
        occurred_at: { type: string, format: date-time }
        title: { type: string }
        description: { type: [string, 'null'] }
        assessment_id: { type: [integer, 'null'] }
        intelligence_object_id: { type: [integer, 'null'] }
        change_id: { type: [integer, 'null'] }
        severity: { type: [string, 'null'] }
    BusinessObjectProfile:
      type: object
      required: [enterprise_object_id, project_id, business_domain, name, object_type, definition, current_status, related_intelligence, evidence, relationships, suggested_questions]
      properties:
        enterprise_object_id: { type: integer }
        project_id: { type: integer }
        business_domain: { type: string }
        name: { type: string }
        object_type: { type: string }
        definition: { type: string }
        aliases: { type: array, items: { type: string } }
        business_importance: { type: [string, 'null'] }
        current_status: { $ref: '#/components/schemas/BusinessHealth' }
        related_intelligence:
          type: array
          items: { $ref: '#/components/schemas/EnterpriseIntelligenceObject' }
        evidence:
          type: array
          items: { $ref: '#/components/schemas/EvidenceReference' }
        relationships:
          type: array
          items:
            type: object
            required: [relationship_type, target_object_id, target_name]
            properties:
              relationship_type: { type: string }
              target_object_id: { type: integer }
              target_name: { type: string }
              explanation: { type: [string, 'null'] }
              confidence: { type: [number, 'null'], minimum: 0, maximum: 1 }
        recent_changes:
          type: array
          items: { $ref: '#/components/schemas/TimelineEvent' }
        suggested_questions: { type: array, items: { type: string } }